backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todo-backend"
version = "1.0.0"
description = "AI-Powered Todo Chatbot Backend"
authors = [{name = "Developer", email = "dev@example.com"}]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sqlmodel>=0.0.16",
    "aiosqlite>=0.19.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "openai>=1.3.7",
    "mcp>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "backend.*"]